+++ packages/api/src/defaultStyleOptions.ts
@@ -313,6 +313,9 @@ const DEFAULT_OPTIONS: Required<StyleOptions> = {
   // Speech recognition
   speechRecognitionContinuous: false,
 
+  partGroupDefaultOpen: true,
+  referenceListDefaultOpen: true,
+
   groupActivitiesBy: ['sender', 'status', 'part'],
 
   // Send box attachment bar